group all features to one folder
